submatview: Move Materializer to submatview package
